+++ controllers/rabbitmqcluster_controller.go
@@ -11,26 +11,17 @@ This product may include a number of subcomponents with separate copyright notic
 package controllers
 
 import (
-	"bufio"
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
-	"reflect"
-	"strings"
-	"time"
-
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/remotecommand"
-
-	"k8s.io/apimachinery/pkg/labels"
-
 	"github.com/rabbitmq/cluster-operator/internal/resource"
 	"github.com/rabbitmq/cluster-operator/internal/status"
 	"k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/record"
+	"reflect"
+	"strings"
 
 	"k8s.io/apimachinery/pkg/types"
 
@@ -55,13 +46,8 @@ var (
 )
 
 const (
-	ownerKey                 = ".metadata.controller"
-	ownerKind                = "RabbitmqCluster"
-	deletionFinalizer        = "deletion.finalizers.rabbitmqclusters.rabbitmq.com"
-	pluginsUpdateAnnotation  = "rabbitmq.com/pluginsUpdatedAt"
-	queueRebalanceAnnotation = "rabbitmq.com/queueRebalanceNeededAt"
-	serverConfAnnotation     = "rabbitmq.com/serverConfUpdatedAt"
-	stsRestartAnnotation     = "rabbitmq.com/lastRestartAt"
+	ownerKey  = ".metadata.controller"
+	ownerKind = "RabbitmqCluster"
 )
 
 // RabbitmqClusterReconciler reconciles a RabbitmqCluster object
@@ -239,314 +225,6 @@ func (r *RabbitmqClusterReconciler) Reconcile(req ctrl.Request) (ctrl.Result, er
 	return ctrl.Result{}, nil
 }
 
-func (r *RabbitmqClusterReconciler) checkTLSSecrets(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqCluster) (ctrl.Result, error) {
-	logger := r.Log
-	secretName := rabbitmqCluster.Spec.TLS.SecretName
-	logger.Info("TLS set, looking for secret", "secret", secretName, "namespace", rabbitmqCluster.Namespace)
-
-	// check if secret exists
-	secret := &corev1.Secret{}
-	if err := r.Get(ctx, types.NamespacedName{Namespace: rabbitmqCluster.Namespace, Name: secretName}, secret); err != nil {
-		r.Recorder.Event(rabbitmqCluster, corev1.EventTypeWarning, "TLSError",
-			fmt.Sprintf("Failed to get TLS secret %v in namespace %v: %v", secretName, rabbitmqCluster.Namespace, err.Error()))
-		return ctrl.Result{}, err
-	}
-	// check if secret has the right keys
-	_, hasTLSKey := secret.Data["tls.key"]
-	_, hasTLSCert := secret.Data["tls.crt"]
-	if !hasTLSCert || !hasTLSKey {
-		r.Recorder.Event(rabbitmqCluster, corev1.EventTypeWarning, "TLSError",
-			fmt.Sprintf("The TLS secret %v in namespace %v must have the fields tls.crt and tls.key", secretName, rabbitmqCluster.Namespace))
-
-		return ctrl.Result{}, errors.NewBadRequest("The TLS secret must have the fields tls.crt and tls.key")
-	}
-
-	// Mutual TLS: check if CA certificate is stored in a separate secret
-	if rabbitmqCluster.MutualTLSEnabled() {
-		if !rabbitmqCluster.SingleTLSSecret() {
-			secretName := rabbitmqCluster.Spec.TLS.CaSecretName
-			logger.Info("mutual TLS set, looking for CA certificate secret", "secret", secretName, "namespace", rabbitmqCluster.Namespace)
-
-			// check if secret exists
-			secret = &corev1.Secret{}
-			if err := r.Get(ctx, types.NamespacedName{Namespace: rabbitmqCluster.Namespace, Name: secretName}, secret); err != nil {
-				r.Recorder.Event(rabbitmqCluster, corev1.EventTypeWarning, "TLSError",
-					fmt.Sprintf("Failed to get CA certificate secret %v in namespace %v: %v", secretName, rabbitmqCluster.Namespace, err.Error()))
-				return ctrl.Result{}, err
-			}
-		}
-		// Mutual TLS: verify that CA certificate is present in secret
-		_, hasCaCert := secret.Data[rabbitmqCluster.Spec.TLS.CaCertName]
-		if !hasCaCert {
-			r.Recorder.Event(rabbitmqCluster, corev1.EventTypeWarning, "TLSError",
-				fmt.Sprintf("The TLS secret %v in namespace %v must have the field %v", rabbitmqCluster.Spec.TLS.CaSecretName, rabbitmqCluster.Namespace, rabbitmqCluster.Spec.TLS.CaCertName))
-
-			return ctrl.Result{}, errors.NewBadRequest(fmt.Sprintf("The TLS secret must have the field %s", rabbitmqCluster.Spec.TLS.CaCertName))
-		}
-	}
-	return ctrl.Result{}, nil
-}
-
-func (r *RabbitmqClusterReconciler) setDefaultUserStatus(ctx context.Context, rmq *rabbitmqv1beta1.RabbitmqCluster) error {
-
-	defaultUserStatus := &rabbitmqv1beta1.RabbitmqClusterDefaultUser{}
-
-	serviceRef := &rabbitmqv1beta1.RabbitmqClusterServiceReference{
-		Name:      rmq.ChildResourceName("client"),
-		Namespace: rmq.Namespace,
-	}
-	defaultUserStatus.ServiceReference = serviceRef
-
-	secretRef := &rabbitmqv1beta1.RabbitmqClusterSecretReference{
-		Name:      rmq.ChildResourceName(resource.DefaultUserSecretName),
-		Namespace: rmq.Namespace,
-		Keys: map[string]string{
-			"username": "username",
-			"password": "password",
-		},
-	}
-	defaultUserStatus.SecretReference = secretRef
-
-	if !reflect.DeepEqual(rmq.Status.DefaultUser, defaultUserStatus) {
-		rmq.Status.DefaultUser = defaultUserStatus
-		if err := r.Status().Update(ctx, rmq); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (r *RabbitmqClusterReconciler) markForQueueRebalance(ctx context.Context, rmq *rabbitmqv1beta1.RabbitmqCluster) error {
-	if rmq.ObjectMeta.Annotations == nil {
-		rmq.ObjectMeta.Annotations = make(map[string]string)
-	}
-
-	if len(rmq.ObjectMeta.Annotations[queueRebalanceAnnotation]) > 0 {
-		return nil
-	}
-
-	rmq.ObjectMeta.Annotations[queueRebalanceAnnotation] = time.Now().Format(time.RFC3339)
-	if err := r.Update(ctx, rmq); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (r *RabbitmqClusterReconciler) runPostDeployStepsIfNeeded(ctx context.Context, rmq *rabbitmqv1beta1.RabbitmqCluster) (requeueAfter time.Duration, err error) {
-	sts, err := r.statefulSet(ctx, rmq)
-	if err != nil {
-		return 0, err
-	}
-	if !allReplicasReadyAndUpdated(sts) {
-		r.Log.Info("not all replicas ready yet; requeuing request to run post deploy steps",
-			"namespace", rmq.Namespace,
-			"name", rmq.Name)
-		return 15 * time.Second, nil
-	}
-
-	// Retrieve the plugins config map, if it exists.
-	pluginsConfig, err := r.configMap(ctx, rmq, rmq.ChildResourceName(resource.PluginsConfigName))
-	if client.IgnoreNotFound(err) != nil {
-		return 0, err
-	}
-	updatedRecently, err := pluginsConfigUpdatedRecently(pluginsConfig)
-	if err != nil {
-		return 0, err
-	}
-	if updatedRecently {
-		// plugins configMap was updated very recently
-		// give StatefulSet controller some time to trigger restart of StatefulSet if necessary
-		// otherwise, there would be race conditions where we exec into containers losing the connection due to pods being terminated
-		r.Log.Info("requeuing request to set plugins on RabbitmqCluster",
-			"namespace", rmq.Namespace,
-			"name", rmq.Name)
-		return 2 * time.Second, nil
-	}
-
-	if pluginsConfig != nil {
-		if err = r.runSetPluginsCommand(ctx, rmq, pluginsConfig); err != nil {
-			return 0, err
-		}
-	}
-
-	// If the cluster has been marked as needing it, run rabbitmq-queues rebalance all
-	if rmq.ObjectMeta.Annotations != nil && len(rmq.ObjectMeta.Annotations[queueRebalanceAnnotation]) > 0 {
-		err = r.runQueueRebalanceCommand(ctx, rmq)
-	}
-	return 0, err
-}
-
-func (r *RabbitmqClusterReconciler) runQueueRebalanceCommand(ctx context.Context, rmq *rabbitmqv1beta1.RabbitmqCluster) error {
-	podName := fmt.Sprintf("%s-0", rmq.ChildResourceName("server"))
-	stdout, stderr, err := r.exec(rmq.Namespace, podName, "rabbitmq", "sh", "-c", "rabbitmq-queues rebalance all")
-	if err != nil {
-		r.Log.Error(err, "failed to run queue rebalance",
-			"namespace", rmq.Namespace,
-			"name", rmq.Name,
-			"pod", podName,
-			"command", "rabbitmq-queues rebalance all",
-			"stdout", stdout,
-			"stderr", stderr)
-		return err
-	}
-	delete(rmq.ObjectMeta.Annotations, queueRebalanceAnnotation)
-	return r.Update(ctx, rmq)
-}
-
-// There are 2 paths how plugins are set:
-// 1. When StatefulSet is (re)started, the up-to-date plugins list (ConfigMap copied by the init container) is read by RabbitMQ nodes during node start up.
-// 2. When the plugins ConfigMap is changed, 'rabbitmq-plugins set' updates the plugins on every node (without the need to re-start the nodes).
-// This method implements the 2nd path.
-func (r *RabbitmqClusterReconciler) runSetPluginsCommand(ctx context.Context, rmq *rabbitmqv1beta1.RabbitmqCluster, configMap *corev1.ConfigMap) error {
-	plugins := resource.NewRabbitmqPlugins(rmq.Spec.Rabbitmq.AdditionalPlugins)
-	for i := int32(0); i < *rmq.Spec.Replicas; i++ {
-		podName := fmt.Sprintf("%s-%d", rmq.ChildResourceName("server"), i)
-		rabbitCommand := fmt.Sprintf("rabbitmq-plugins set %s", plugins.AsString(" "))
-		stdout, stderr, err := r.exec(rmq.Namespace, podName, "rabbitmq", "sh", "-c", rabbitCommand)
-		if err != nil {
-			r.Log.Error(err, "failed to set plugins",
-				"namespace", rmq.Namespace,
-				"name", rmq.Name,
-				"pod", podName,
-				"command", rabbitCommand,
-				"stdout", stdout,
-				"stderr", stderr)
-			return err
-		}
-	}
-	r.Log.Info("successfully set plugins on RabbitmqCluster",
-		"namespace", rmq.Namespace,
-		"name", rmq.Name)
-
-	delete(configMap.Annotations, pluginsUpdateAnnotation)
-	if err := r.Update(ctx, configMap); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Adds an arbitrary annotation to the sts PodTemplate to trigger a sts restart
-// it compares annotation "rabbitmq.com/serverConfUpdatedAt" from server-conf configMap and annotation "rabbitmq.com/lastRestartAt" from sts
-// to determine whether to restart sts
-func (r *RabbitmqClusterReconciler) restartStatefulSetIfNeeded(ctx context.Context, rmq *rabbitmqv1beta1.RabbitmqCluster) (time.Duration, error) {
-	serverConf, err := r.configMap(ctx, rmq, rmq.ChildResourceName(resource.ServerConfigMapName))
-	if err != nil {
-		// requeue request after 10s if unable to find server-conf configmap, else return the error
-		return 10 * time.Second, client.IgnoreNotFound(err)
-	}
-
-	serverConfigUpdatedAt, ok := serverConf.Annotations[serverConfAnnotation]
-	if !ok {
-		// server-conf configmap hasn't been updated; no need to restart sts
-		return 0, nil
-	}
-
-	sts, err := r.statefulSet(ctx, rmq)
-	if err != nil {
-		// requeue request after 10s if unable to find sts, else return the error
-		return 10 * time.Second, client.IgnoreNotFound(err)
-	}
-
-	stsRestartedAt, ok := sts.Spec.Template.ObjectMeta.Annotations[stsRestartAnnotation]
-	if ok && stsRestartedAt > serverConfigUpdatedAt {
-		// sts was updated after the last server-conf configmap update; no need to restart sts
-		return 0, nil
-	}
-
-	if err := clientretry.RetryOnConflict(clientretry.DefaultRetry, func() error {
-		sts := &appsv1.StatefulSet{ObjectMeta: metav1.ObjectMeta{Name: rmq.ChildResourceName("server"), Namespace: rmq.Namespace}}
-		if err := r.Get(ctx, types.NamespacedName{Name: sts.Name, Namespace: sts.Namespace}, sts); err != nil {
-			return err
-		}
-		if sts.Spec.Template.ObjectMeta.Annotations == nil {
-			sts.Spec.Template.ObjectMeta.Annotations = make(map[string]string)
-		}
-		sts.Spec.Template.ObjectMeta.Annotations[stsRestartAnnotation] = time.Now().Format(time.RFC3339)
-		return r.Update(ctx, sts)
-	}); err != nil {
-		msg := fmt.Sprintf("failed to restart StatefulSet %s of Namespace %s; rabbitmq.conf configuration may be outdated", rmq.ChildResourceName("server"), rmq.Namespace)
-		r.Log.Error(err, msg)
-		r.Recorder.Event(rmq, corev1.EventTypeWarning, "FailedUpdate", msg)
-		// failed to restart sts; return error to requeue request
-		return 0, err
-	}
-
-	msg := fmt.Sprintf("restarted StatefulSet %s of Namespace %s", rmq.ChildResourceName("server"), rmq.Namespace)
-	r.Log.Info(msg)
-	r.Recorder.Event(rmq, corev1.EventTypeNormal, "SuccessfulUpdate", msg)
-
-	return 0, nil
-}
-
-func (r *RabbitmqClusterReconciler) statefulSet(ctx context.Context, rmq *rabbitmqv1beta1.RabbitmqCluster) (*appsv1.StatefulSet, error) {
-	sts := &appsv1.StatefulSet{}
-	if err := r.Get(ctx, types.NamespacedName{Name: rmq.ChildResourceName("server"), Namespace: rmq.Namespace}, sts); err != nil {
-		return nil, err
-	}
-	return sts, nil
-}
-
-func (r *RabbitmqClusterReconciler) configMap(ctx context.Context, rmq *rabbitmqv1beta1.RabbitmqCluster, name string) (*corev1.ConfigMap, error) {
-	configMap := &corev1.ConfigMap{}
-	if err := r.Get(ctx, types.NamespacedName{Namespace: rmq.Namespace, Name: name}, configMap); err != nil {
-		return nil, err
-	}
-	return configMap, nil
-}
-
-// Annotates the plugins ConfigMap or the server-conf ConfigMap
-// annotations later used to indicate whether to call 'rabbitmq-plugins set' or to restart the sts
-func (r *RabbitmqClusterReconciler) annotateConfigMapIfUpdated(ctx context.Context, builder resource.ResourceBuilder, operationResult controllerutil.OperationResult, rmq *rabbitmqv1beta1.RabbitmqCluster) error {
-	if operationResult != controllerutil.OperationResultUpdated {
-		return nil
-	}
-
-	var configMap, annotationKey string
-	switch builder.(type) {
-	case *resource.RabbitmqPluginsConfigMapBuilder:
-		configMap = rmq.ChildResourceName(resource.PluginsConfigName)
-		annotationKey = pluginsUpdateAnnotation
-	case *resource.ServerConfigMapBuilder:
-		configMap = rmq.ChildResourceName(resource.ServerConfigMapName)
-		annotationKey = serverConfAnnotation
-	default:
-		return nil
-	}
-
-	if err := r.annotateConfigMap(ctx, rmq.Namespace, configMap, annotationKey, time.Now().Format(time.RFC3339)); err != nil {
-		msg := fmt.Sprintf("Failed to annotate ConfigMap %s of Namespace %s; %s may be outdated", configMap, rmq.Namespace, rmq.Name)
-		r.Log.Error(err, msg)
-		r.Recorder.Event(rmq, corev1.EventTypeWarning, "FailedUpdate", msg)
-		return err
-	}
-
-	r.Log.Info("successfully annotated", "ConfigMap", configMap, "Namespace", rmq.Namespace)
-	return nil
-}
-
-func (r *RabbitmqClusterReconciler) annotateConfigMap(ctx context.Context, namespace, name, key, value string) error {
-	if retryOnConflictErr := clientretry.RetryOnConflict(clientretry.DefaultRetry, func() error {
-		configMap := corev1.ConfigMap{}
-		if err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: name}, &configMap); err != nil {
-			return client.IgnoreNotFound(err)
-		}
-		if configMap.Annotations == nil {
-			configMap.Annotations = make(map[string]string)
-		}
-		configMap.Annotations[key] = value
-		return r.Update(ctx, &configMap)
-	}); retryOnConflictErr != nil {
-		return retryOnConflictErr
-	}
-	return nil
-}
-
-func (r *RabbitmqClusterReconciler) exec(namespace, podName, containerName string, command ...string) (string, string, error) {
-	return r.PodExecutor.Exec(r.Clientset, r.ClusterConfig, namespace, podName, containerName, command...)
-}
-
 // logAndRecordOperationResult - helper function to log and record events with message and error
 // it logs and records 'updated' and 'created' OperationResult, and ignores OperationResult 'unchanged'
 func (r *RabbitmqClusterReconciler) logAndRecordOperationResult(rmq runtime.Object, resource runtime.Object, operationResult controllerutil.OperationResult, err error) {
@@ -573,84 +251,6 @@ func (r *RabbitmqClusterReconciler) logAndRecordOperationResult(rmq runtime.Obje
 		r.Log.Error(err, msg)
 		r.Recorder.Event(rmq, corev1.EventTypeWarning, fmt.Sprintf("Failed%s", strings.Title(operation)), msg)
 	}
-}
-
-func (r *RabbitmqClusterReconciler) prepareForDeletion(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqCluster) error {
-	if controllerutil.ContainsFinalizer(rabbitmqCluster, deletionFinalizer) {
-		if err := clientretry.RetryOnConflict(clientretry.DefaultRetry, func() error {
-			sts := &appsv1.StatefulSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      rabbitmqCluster.ChildResourceName("server"),
-					Namespace: rabbitmqCluster.Namespace,
-				},
-			}
-			// Add label on all Pods to be picked up in pre-stop hook via Downward API
-			if err := r.addRabbitmqDeletionLabel(ctx, rabbitmqCluster); err != nil {
-				return fmt.Errorf("failed to add deletion markers to RabbitmqCluster Pods: %s", err.Error())
-			}
-			// Delete StatefulSet immediately after changing pod labels to minimize risk of them respawning.
-			// There is a window where the StatefulSet could respawn Pods without the deletion label in this order.
-			// But we can't delete it before because the DownwardAPI doesn't update once a Pod enters Terminating.
-			if err := r.Client.Delete(ctx, sts); client.IgnoreNotFound(err) != nil {
-				return fmt.Errorf("cannot delete StatefulSet: %s", err.Error())
-			}
-
-			return nil
-		}); err != nil {
-			r.Log.Error(err, "RabbitmqCluster deletion")
-		}
-
-		if err := r.removeFinalizer(ctx, rabbitmqCluster); err != nil {
-			r.Log.Error(err, "Failed to remove finalizer for deletion")
-			return err
-		}
-	}
-	return nil
-}
-
-func (r *RabbitmqClusterReconciler) removeFinalizer(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqCluster) error {
-	controllerutil.RemoveFinalizer(rabbitmqCluster, deletionFinalizer)
-	if err := r.Client.Update(ctx, rabbitmqCluster); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (r *RabbitmqClusterReconciler) addRabbitmqDeletionLabel(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqCluster) error {
-	pods := &corev1.PodList{}
-	selector, err := labels.Parse(fmt.Sprintf("app.kubernetes.io/name=%s", rabbitmqCluster.Name))
-	if err != nil {
-		return err
-	}
-	listOptions := client.ListOptions{
-		LabelSelector: selector,
-	}
-
-	if err := r.Client.List(ctx, pods, &listOptions); err != nil {
-		return err
-	}
-
-	for i := 0; i < len(pods.Items); i++ {
-		pod := &pods.Items[i]
-		pod.Labels[resource.DeletionMarker] = "true"
-		if err := r.Client.Update(ctx, pod); client.IgnoreNotFound(err) != nil {
-			return fmt.Errorf("cannot Update Pod %s in Namespace %s: %s", pod.Name, pod.Namespace, err.Error())
-		}
-	}
-
-	return nil
-}
-
-// addFinalizerIfNeeded adds a deletion finalizer if the RabbitmqCluster does not have one yet and is not marked for deletion
-func (r *RabbitmqClusterReconciler) addFinalizerIfNeeded(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqCluster) error {
-	if rabbitmqCluster.ObjectMeta.DeletionTimestamp.IsZero() && !controllerutil.ContainsFinalizer(rabbitmqCluster, deletionFinalizer) {
-		controllerutil.AddFinalizer(rabbitmqCluster, deletionFinalizer)
-		if err := r.Client.Update(ctx, rabbitmqCluster); err != nil {
-			return err
-		}
-	}
-	return nil
 }
 
 func (r *RabbitmqClusterReconciler) getChildResources(ctx context.Context, rmq rabbitmqv1beta1.RabbitmqCluster) ([]runtime.Object, error) {
@@ -737,81 +337,4 @@ func validateAndGetOwner(owner *metav1.OwnerReference) []string {
 		return nil
 	}
 	return []string{owner.Name}
-}
-
-func allReplicasReadyAndUpdated(sts *appsv1.StatefulSet) bool {
-	return sts.Status.ReadyReplicas == *sts.Spec.Replicas && !statefulSetBeingUpdated(sts)
-}
-
-func statefulSetBeingUpdated(sts *appsv1.StatefulSet) bool {
-	return sts.Status.CurrentRevision != sts.Status.UpdateRevision
-}
-
-func statefulSetNeedsQueueRebalance(sts *appsv1.StatefulSet, rmq *rabbitmqv1beta1.RabbitmqCluster) bool {
-	return statefulSetBeingUpdated(sts) &&
-		!rmq.Spec.SkipPostDeploySteps &&
-		*rmq.Spec.Replicas > 1
-}
-
-func pluginsConfigUpdatedRecently(cfg *corev1.ConfigMap) (bool, error) {
-	if cfg == nil {
-		return false, nil
-	}
-	pluginsUpdatedAt, ok := cfg.Annotations[pluginsUpdateAnnotation]
-	if !ok {
-		return false, nil // plugins configMap was not updated
-	}
-
-	annotationTime, err := time.Parse(time.RFC3339, pluginsUpdatedAt)
-	if err != nil {
-		return false, err
-	}
-	return time.Since(annotationTime).Seconds() < 2, nil
-}
-
-type PodExecutor interface {
-	Exec(clientset *kubernetes.Clientset, clusterConfig *rest.Config, namespace, podName, containerName string, command ...string) (string, string, error)
-}
-
-func NewPodExecutor() PodExecutor { return &rabbitmqPodExecutor{} }
-
-type rabbitmqPodExecutor struct{}
-
-func (p *rabbitmqPodExecutor) Exec(clientset *kubernetes.Clientset, clusterConfig *rest.Config, namespace, podName, containerName string, command ...string) (string, string, error) {
-	request := clientset.CoreV1().RESTClient().
-		Post().
-		Resource("pods").
-		Name(podName).
-		Namespace(namespace).
-		SubResource("exec").
-		VersionedParams(&corev1.PodExecOptions{
-			Container: containerName,
-			Command:   command,
-			Stdout:    true,
-			Stderr:    true,
-			Stdin:     false,
-		}, scheme.ParameterCodec)
-
-	exec, err := remotecommand.NewSPDYExecutor(clusterConfig, "POST", request.URL())
-	if err != nil {
-		return "", "", err
-	}
-
-	stdOut := bytes.Buffer{}
-	stdErr := bytes.Buffer{}
-
-	err = exec.Stream(remotecommand.StreamOptions{
-		Stdout: bufio.NewWriter(&stdOut),
-		Stderr: bufio.NewWriter(&stdErr),
-		Stdin:  nil,
-		Tty:    false,
-	})
-	if err != nil {
-		return stdOut.String(), stdErr.String(), err
-	}
-	if stdErr.Len() > 0 {
-		return stdOut.String(), stdErr.String(), fmt.Errorf("%v", stdErr)
-	}
-
-	return stdOut.String(), "", nil
 }
