+++ controllers/rabbitmqcluster_controller_test.go
@@ -18,10 +18,9 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 
 	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/ginkgo/extensions/table"
+
 	. "github.com/onsi/gomega"
 	rabbitmqv1beta1 "github.com/rabbitmq/cluster-operator/api/v1beta1"
-	"github.com/rabbitmq/cluster-operator/internal/resource"
 	"github.com/rabbitmq/cluster-operator/internal/status"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -29,7 +28,6 @@ import (
 	k8sresource "k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/util/retry"
 	runtimeClient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -43,24 +41,7 @@ var _ = Describe("RabbitmqClusterController", func() {
 	var (
 		cluster          *rabbitmqv1beta1.RabbitmqCluster
 		one              int32 = 1
-		three            int32 = 3
 		defaultNamespace       = "default"
-		ctx                    = context.Background()
-		updateWithRetry        = func(cr *rabbitmqv1beta1.RabbitmqCluster, mutateFn func(r *rabbitmqv1beta1.RabbitmqCluster)) error {
-			return retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				objKey, err := runtimeClient.ObjectKeyFromObject(cr)
-				if err != nil {
-					return err
-				}
-
-				if err := client.Get(ctx, objKey, cr); err != nil {
-					return err
-				}
-
-				mutateFn(cr)
-				return client.Update(ctx, cr)
-			})
-		}
 	)
 
 	Context("default settings", func() {
@@ -178,240 +159,7 @@ var _ = Describe("RabbitmqClusterController", func() {
 				Expect(allEventMsgs).To(ContainSubstring("created resource %s of Type *v1.RoleBinding", cluster.ChildResourceName("server")))
 			})
 
-			By("adding the deletion finalizer", func() {
-				rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-				Eventually(func() string {
-					err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-					if err != nil {
-						return ""
-					}
-					if len(rmq.Finalizers) > 0 {
-						return rmq.Finalizers[0]
-					}
-
-					return ""
-				}, 5).Should(Equal("deletion.finalizers.rabbitmqclusters.rabbitmq.com"))
-			})
-
-			By("setting the default-user secret details in the custom resource status", func() {
-				rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-				secretRef := &rabbitmqv1beta1.RabbitmqClusterSecretReference{}
-				Eventually(func() *rabbitmqv1beta1.RabbitmqClusterSecretReference {
-					err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-					if err != nil {
-						return nil
-					}
-
-					if rmq.Status.DefaultUser != nil && rmq.Status.DefaultUser.SecretReference != nil {
-						secretRef = rmq.Status.DefaultUser.SecretReference
-						return secretRef
-					}
-
-					return nil
-				}, 5).ShouldNot(BeNil())
-
-				Expect(secretRef.Name).To(Equal(rmq.ChildResourceName(resource.DefaultUserSecretName)))
-				Expect(secretRef.Namespace).To(Equal(rmq.Namespace))
-				Expect(secretRef.Keys).To(HaveKeyWithValue("username", "username"))
-				Expect(secretRef.Keys).To(HaveKeyWithValue("password", "password"))
-			})
-
-			By("setting the client service details in the custom resource status", func() {
-				rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-				serviceRef := &rabbitmqv1beta1.RabbitmqClusterServiceReference{}
-				Eventually(func() *rabbitmqv1beta1.RabbitmqClusterServiceReference {
-					err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-					if err != nil {
-						return nil
-					}
-
-					if rmq.Status.DefaultUser != nil && rmq.Status.DefaultUser.ServiceReference != nil {
-						serviceRef = rmq.Status.DefaultUser.ServiceReference
-						return serviceRef
-					}
-
-					return nil
-				}, 5).ShouldNot(BeNil())
-
-				Expect(serviceRef.Name).To(Equal(rmq.ChildResourceName("client")))
-				Expect(serviceRef.Namespace).To(Equal(rmq.Namespace))
-			})
 		})
-	})
-
-	Context("Mutual TLS", func() {
-		AfterEach(func() {
-			Expect(client.Delete(ctx, cluster)).To(Succeed())
-			Eventually(func() bool {
-				rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-				err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-				return apierrors.IsNotFound(err)
-			}, 5).Should(BeTrue())
-		})
-
-		Context("Mutual TLS with single secret", func() {
-			It("Deploys successfully", func() {
-				tlsSecretWithCACert(ctx, "tls-secret", defaultNamespace, "caCERT")
-				tlsSpec := rabbitmqv1beta1.TLSSpec{
-					SecretName:   "tls-secret",
-					CaSecretName: "tls-secret",
-					CaCertName:   "caCERT",
-				}
-				cluster = rabbitmqClusterWithTLS(ctx, "mutual-tls-success", defaultNamespace, tlsSpec)
-				waitForClusterCreation(ctx, cluster, client)
-
-				sts, err := clientSet.AppsV1().StatefulSets(cluster.Namespace).Get(ctx, cluster.ChildResourceName("server"), metav1.GetOptions{})
-				Expect(err).NotTo(HaveOccurred())
-				volumeMount := corev1.VolumeMount{
-					Name:      "rabbitmq-tls",
-					MountPath: "/etc/rabbitmq-tls/caCERT",
-					SubPath:   "caCERT",
-					ReadOnly:  true,
-				}
-				Expect(sts.Spec.Template.Spec.Containers[0].VolumeMounts).To(ContainElement(volumeMount))
-			})
-
-			It("Does not deploy if the cert name does not match the contents of the secret", func() {
-				tlsSecretWithCACert(ctx, "tls-secret-missing", defaultNamespace, "ca.c")
-				tlsSpec := rabbitmqv1beta1.TLSSpec{
-					SecretName:   "tls-secret-missing",
-					CaSecretName: "tls-secret-missing",
-					CaCertName:   "ca.crt",
-				}
-				cluster = rabbitmqClusterWithTLS(ctx, "tls-secret-missing", defaultNamespace, tlsSpec)
-
-				verifyError(ctx, cluster, fmt.Sprintf("The TLS secret tls-secret-missing in namespace %s must have the field ca.crt", defaultNamespace))
-			})
-		})
-
-		Context("Mutual TLS with a seperate CA certificate secret", func() {
-			It("Does not deploy the RabbitmqCluster, and retries every 10 seconds", func() {
-				tlsSecretWithoutCACert(ctx, "rabbitmq-tls-secret-does-not-exist", defaultNamespace)
-
-				tlsSpec := rabbitmqv1beta1.TLSSpec{
-					SecretName:   "rabbitmq-tls-secret-does-not-exist",
-					CaSecretName: "ca-cert-secret",
-					CaCertName:   "ca.crt",
-				}
-				cluster = rabbitmqClusterWithTLS(ctx, "rabbitmq-tls-secret-does-not-exist", defaultNamespace, tlsSpec)
-				verifyError(ctx, cluster, "Failed to get CA certificate secret")
-
-				_, err := clientSet.AppsV1().StatefulSets(cluster.Namespace).Get(ctx, cluster.ChildResourceName("server"), metav1.GetOptions{})
-				Expect(err).To(HaveOccurred())
-
-				// create missing secret
-				caData := map[string]string{
-					"ca.crt": "this is a ca cert",
-				}
-				_, err = createSecret(ctx, "ca-cert-secret", defaultNamespace, caData)
-				Expect(err).NotTo(HaveOccurred())
-
-				waitForClusterCreation(ctx, cluster, client)
-				statefulSet(ctx, cluster)
-			})
-
-			It("Does not deploy if the cert name does not match the contents of the secret", func() {
-				tlsSecretWithoutCACert(ctx, "tls-secret", defaultNamespace)
-				caData := map[string]string{
-					"cacrt": "this is a ca cert",
-				}
-
-				_, err := createSecret(ctx, "ca-cert-secret-invalid", defaultNamespace, caData)
-				if !apierrors.IsAlreadyExists(err) {
-					Expect(err).NotTo(HaveOccurred())
-				}
-
-				tlsSpec := rabbitmqv1beta1.TLSSpec{
-					SecretName:   "tls-secret",
-					CaSecretName: "ca-cert-secret-invalid",
-					CaCertName:   "ca.crt",
-				}
-				cluster = rabbitmqClusterWithTLS(ctx, "rabbitmq-mutual-tls-missing", defaultNamespace, tlsSpec)
-				verifyError(ctx, cluster, fmt.Sprintf("The TLS secret ca-cert-secret-invalid in namespace %s must have the field ca.crt", defaultNamespace))
-			})
-		})
-	})
-
-	Context("TLS set on the instance", func() {
-		AfterEach(func() {
-			Expect(client.Delete(ctx, cluster)).To(Succeed())
-			Eventually(func() bool {
-				rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-				err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-				return apierrors.IsNotFound(err)
-			}, 5).Should(BeTrue())
-		})
-
-		BeforeEach(func() {
-			tlsSecretWithoutCACert(ctx, "tls-secret", defaultNamespace)
-		})
-
-		It("Deploys successfully", func() {
-			cluster = &rabbitmqv1beta1.RabbitmqCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "rabbitmq-tls",
-					Namespace: defaultNamespace,
-				},
-				Spec: rabbitmqv1beta1.RabbitmqClusterSpec{
-					TLS: rabbitmqv1beta1.TLSSpec{
-						SecretName: "tls-secret",
-					},
-				},
-			}
-
-			Expect(client.Create(ctx, cluster)).To(Succeed())
-			waitForClusterCreation(ctx, cluster, client)
-		})
-
-		When("the TLS secret does not have the expected keys - tls.crt, or tls.key", func() {
-			BeforeEach(func() {
-				secretData := map[string]string{
-					"somekey": "someval",
-					"tls.key": "this is a tls key",
-				}
-				_, err := createSecret(ctx, "rabbitmq-tls-malformed", defaultNamespace, secretData)
-
-				if !apierrors.IsAlreadyExists(err) {
-					Expect(err).NotTo(HaveOccurred())
-				}
-
-				tlsSpec := rabbitmqv1beta1.TLSSpec{
-					SecretName: "rabbitmq-tls-malformed",
-				}
-				cluster = rabbitmqClusterWithTLS(ctx, "rabbitmq-tls-malformed", defaultNamespace, tlsSpec)
-			})
-
-			It("fails to deploy the RabbitmqCluster", func() {
-				verifyError(ctx, cluster, fmt.Sprintf("The TLS secret rabbitmq-tls-malformed in namespace %s must have the fields tls.crt and tls.key", defaultNamespace))
-			})
-		})
-
-		When("the TLS secret does not exist", func() {
-			It("fails to deploy the RabbitmqCluster until the secret is detected", func() {
-
-				tlsSpec := rabbitmqv1beta1.TLSSpec{
-					SecretName: "tls-secret-does-not-exist",
-				}
-				cluster = rabbitmqClusterWithTLS(ctx, "rabbitmq-tls-secret-does-not-exist", defaultNamespace, tlsSpec)
-
-				verifyError(ctx, cluster, "Failed to get TLS secret")
-
-				_, err := clientSet.AppsV1().StatefulSets(cluster.Namespace).Get(ctx, cluster.ChildResourceName("server"), metav1.GetOptions{})
-				Expect(err).To(HaveOccurred())
-
-				// create missing secret
-				secretData := map[string]string{
-					"tls.crt": "this is a tls cert",
-					"tls.key": "this is a tls key",
-				}
-				_, err = createSecret(ctx, "tls-secret-does-not-exist", defaultNamespace, secretData)
-				Expect(err).NotTo(HaveOccurred())
-
-				waitForClusterCreation(ctx, cluster, client)
-				statefulSet(ctx, cluster)
-			})
-		})
-
 	})
 
 	Context("Annotations set on the instance", func() {
@@ -1379,297 +1127,6 @@ var _ = Describe("RabbitmqClusterController", func() {
 		})
 	})
 
-	Context("Cluster post deploy steps", func() {
-		var annotations map[string]string
-
-		BeforeEach(func() {
-			annotations = map[string]string{}
-		})
-
-		AfterEach(func() {
-			Expect(client.Delete(ctx, cluster)).To(Succeed())
-			waitForClusterDeletion(ctx, cluster, client)
-		})
-
-		When("the cluster is configured to run post-deploy steps", func() {
-			BeforeEach(func() {
-				cluster = &rabbitmqv1beta1.RabbitmqCluster{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "rabbitmq-three",
-						Namespace: defaultNamespace,
-					},
-					Spec: rabbitmqv1beta1.RabbitmqClusterSpec{
-						Replicas: &three,
-					},
-				}
-				Expect(client.Create(ctx, cluster)).To(Succeed())
-				waitForClusterCreation(ctx, cluster, client)
-			})
-			When("the cluster is updated", func() {
-				var sts *appsv1.StatefulSet
-
-				BeforeEach(func() {
-					sts = statefulSet(ctx, cluster)
-					sts.Status.Replicas = 3
-					sts.Status.CurrentReplicas = 2
-					sts.Status.CurrentRevision = "some-old-revision"
-					sts.Status.UpdatedReplicas = 1
-					sts.Status.UpdateRevision = "some-new-revision"
-
-					statusWriter := client.Status()
-					err := statusWriter.Update(ctx, sts)
-					Expect(err).NotTo(HaveOccurred())
-				})
-
-				It("triggers the controller to run rabbitmq-queues rebalance all", func() {
-					By("setting an annotation on the CR", func() {
-						Eventually(func() map[string]string {
-							rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-							err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-							Expect(err).To(BeNil())
-							annotations = rmq.ObjectMeta.Annotations
-							return annotations
-						}, 5).Should(HaveKey("rabbitmq.com/queueRebalanceNeededAt"))
-						_, err := time.Parse(time.RFC3339, annotations["rabbitmq.com/queueRebalanceNeededAt"])
-						Expect(err).NotTo(HaveOccurred(), "Annotation rabbitmq.com/queueRebalanceNeededAt was not a valid RFC3339 timestamp")
-					})
-
-					By("not removing the annotation when all replicas are updated but not yet ready", func() {
-						sts.Status.CurrentReplicas = 3
-						sts.Status.CurrentRevision = "some-new-revision"
-						sts.Status.UpdatedReplicas = 3
-						sts.Status.UpdateRevision = "some-new-revision"
-						sts.Status.ReadyReplicas = 2
-						statusWriter := client.Status()
-						err := statusWriter.Update(ctx, sts)
-						Expect(err).NotTo(HaveOccurred())
-						Eventually(func() map[string]string {
-							rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-							err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-							Expect(err).To(BeNil())
-							annotations = rmq.ObjectMeta.Annotations
-							return annotations
-						}, 5).Should(HaveKey("rabbitmq.com/queueRebalanceNeededAt"))
-						Expect(fakeExecutor.ExecutedCommands()).NotTo(ContainElement(command{"sh", "-c", "rabbitmq-queues rebalance all"}))
-						_, err = time.Parse(time.RFC3339, annotations["rabbitmq.com/queueRebalanceNeededAt"])
-						Expect(err).NotTo(HaveOccurred(), "Annotation rabbitmq.com/queueRebalanceNeededAt was not a valid RFC3339 timestamp")
-					})
-
-					By("removing the annotation once all Pods are up, and triggering the queue rebalance", func() {
-						sts.Status.ReadyReplicas = 3
-						statusWriter := client.Status()
-						err := statusWriter.Update(ctx, sts)
-						Expect(err).NotTo(HaveOccurred())
-						Eventually(func() map[string]string {
-							rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-							err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-							Expect(err).To(BeNil())
-							return rmq.ObjectMeta.Annotations
-						}, 5).ShouldNot(HaveKey("rabbitmq.com/queueRebalanceNeededAt"))
-						Expect(fakeExecutor.ExecutedCommands()).To(ContainElement(command{"sh", "-c", "rabbitmq-queues rebalance all"}))
-					})
-				})
-
-			})
-		})
-
-		When("the cluster is not configured to run post-deploy steps", func() {
-			BeforeEach(func() {
-				cluster = &rabbitmqv1beta1.RabbitmqCluster{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "rabbitmq-three-no-post-deploy",
-						Namespace: defaultNamespace,
-					},
-					Spec: rabbitmqv1beta1.RabbitmqClusterSpec{
-						Replicas:            &three,
-						SkipPostDeploySteps: true,
-					},
-				}
-				Expect(client.Create(ctx, cluster)).To(Succeed())
-				waitForClusterCreation(ctx, cluster, client)
-			})
-			When("the cluster is updated", func() {
-				var sts *appsv1.StatefulSet
-
-				BeforeEach(func() {
-					sts = statefulSet(ctx, cluster)
-					sts.Status.Replicas = 3
-					sts.Status.CurrentReplicas = 2
-					sts.Status.CurrentRevision = "some-old-revision"
-					sts.Status.UpdatedReplicas = 1
-					sts.Status.UpdateRevision = "some-new-revision"
-
-					statusWriter := client.Status()
-					err := statusWriter.Update(ctx, sts)
-					Expect(err).NotTo(HaveOccurred())
-				})
-
-				It("does not trigger the controller to run rabbitmq-queues rebalance all", func() {
-					By("never setting the annotation on the CR", func() {
-						Consistently(func() map[string]string {
-							rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-							err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-							Expect(err).To(BeNil())
-							return rmq.ObjectMeta.Annotations
-						}, 5).ShouldNot(HaveKey("rabbitmq.com/queueRebalanceNeededAt"))
-					})
-
-					By("not running the rebalance command once all nodes are up", func() {
-						sts.Status.CurrentReplicas = 3
-						sts.Status.CurrentRevision = "some-new-revision"
-						sts.Status.UpdatedReplicas = 3
-						sts.Status.UpdateRevision = "some-new-revision"
-						sts.Status.ReadyReplicas = 3
-						statusWriter := client.Status()
-						err := statusWriter.Update(ctx, sts)
-						Expect(err).NotTo(HaveOccurred())
-						Consistently(func() map[string]string {
-							rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-							err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-							Expect(err).To(BeNil())
-							return rmq.ObjectMeta.Annotations
-						}, 5).ShouldNot(HaveKey("rabbitmq.com/queueRebalanceNeededAt"))
-						Expect(fakeExecutor.ExecutedCommands()).NotTo(ContainElement(command{"sh", "-c", "rabbitmq-queues rebalance all"}))
-					})
-
-				})
-
-			})
-		})
-
-		When("the cluster is a single node cluster", func() {
-			BeforeEach(func() {
-				cluster = &rabbitmqv1beta1.RabbitmqCluster{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "rabbitmq-one-rebalance",
-						Namespace: defaultNamespace,
-					},
-					Spec: rabbitmqv1beta1.RabbitmqClusterSpec{
-						Replicas:            &one,
-						SkipPostDeploySteps: false,
-					},
-				}
-				Expect(client.Create(ctx, cluster)).To(Succeed())
-				waitForClusterCreation(ctx, cluster, client)
-			})
-			When("the cluster is updated", func() {
-				var sts *appsv1.StatefulSet
-
-				BeforeEach(func() {
-					sts = statefulSet(ctx, cluster)
-					sts.Status.Replicas = 1
-					sts.Status.CurrentReplicas = 1
-					sts.Status.CurrentRevision = "some-old-revision"
-					sts.Status.UpdatedReplicas = 0
-					sts.Status.UpdateRevision = "some-new-revision"
-					sts.Status.ReadyReplicas = 0
-
-					statusWriter := client.Status()
-					err := statusWriter.Update(ctx, sts)
-					Expect(err).NotTo(HaveOccurred())
-				})
-
-				It("does not trigger the controller to run rabbitmq-queues rebalance all", func() {
-					By("never setting the annotation on the CR", func() {
-						Consistently(func() map[string]string {
-							rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-							err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-							Expect(err).To(BeNil())
-							return rmq.ObjectMeta.Annotations
-						}, 5).ShouldNot(HaveKey("rabbitmq.com/queueRebalanceNeededAt"))
-					})
-
-					By("not running the rebalance command once all nodes are up", func() {
-						sts.Status.CurrentReplicas = 1
-						sts.Status.CurrentRevision = "some-new-revision"
-						sts.Status.UpdatedReplicas = 1
-						sts.Status.UpdateRevision = "some-new-revision"
-						sts.Status.ReadyReplicas = 1
-						statusWriter := client.Status()
-						err := statusWriter.Update(ctx, sts)
-						Expect(err).NotTo(HaveOccurred())
-						Consistently(func() map[string]string {
-							rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-							err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-							Expect(err).To(BeNil())
-							return rmq.ObjectMeta.Annotations
-						}, 5).ShouldNot(HaveKey("rabbitmq.com/queueRebalanceNeededAt"))
-						Expect(fakeExecutor.ExecutedCommands()).NotTo(ContainElement(command{"sh", "-c", "rabbitmq-queues rebalance all"}))
-					})
-
-				})
-
-			})
-		})
-
-	})
-
-	DescribeTable("Server configurations updates",
-		func(testCase string) {
-			// create rabbitmqcluster
-			cluster = &rabbitmqv1beta1.RabbitmqCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "rabbitmq-" + testCase,
-					Namespace: defaultNamespace,
-				},
-			}
-			Expect(client.Create(ctx, cluster)).To(Succeed())
-			waitForClusterCreation(ctx, cluster, client)
-
-			// ensure that configMap and statefulSet does not have annotations set when configurations haven't changed
-			configMap, err := clientSet.CoreV1().ConfigMaps(cluster.Namespace).Get(ctx, cluster.ChildResourceName("server-conf"), metav1.GetOptions{})
-			Expect(err).To(Not(HaveOccurred()))
-			Expect(configMap.Annotations).ShouldNot(HaveKey("rabbitmq.com/serverConfUpdatedAt"))
-
-			sts, err := clientSet.AppsV1().StatefulSets(cluster.Namespace).Get(ctx, cluster.ChildResourceName("server"), metav1.GetOptions{})
-			Expect(err).To(Not(HaveOccurred()))
-			Expect(sts.Annotations).ShouldNot(HaveKey("rabbitmq.com/lastRestartAt"))
-
-			// update rabbitmq server configurations
-			Expect(updateWithRetry(cluster, func(r *rabbitmqv1beta1.RabbitmqCluster) {
-				if testCase == "additional-config" {
-					r.Spec.Rabbitmq.AdditionalConfig = "test_config=0"
-				}
-				if testCase == "advanced-config" {
-					r.Spec.Rabbitmq.AdvancedConfig = "sample-advanced-config."
-				}
-				if testCase == "env-config" {
-					r.Spec.Rabbitmq.EnvConfig = "some-env-variable"
-				}
-			})).To(Succeed())
-
-			By("annotating the server-conf ConfigMap")
-			// ensure annotations from the server-conf ConfigMap
-			var annotations map[string]string
-			Eventually(func() map[string]string {
-				configMap, err := clientSet.CoreV1().ConfigMaps(cluster.Namespace).Get(ctx, cluster.ChildResourceName("server-conf"), metav1.GetOptions{})
-				Expect(err).To(Not(HaveOccurred()))
-				annotations = configMap.Annotations
-				return annotations
-			}, 5).Should(HaveKey("rabbitmq.com/serverConfUpdatedAt"))
-			_, err = time.Parse(time.RFC3339, annotations["rabbitmq.com/serverConfUpdatedAt"])
-			Expect(err).NotTo(HaveOccurred(), "Annotation rabbitmq.com/serverConfUpdatedAt was not a valid RFC3339 timestamp")
-
-			By("annotating the sts podTemplate")
-			// ensure statefulSet annotations
-			Eventually(func() map[string]string {
-				sts, err := clientSet.AppsV1().StatefulSets(cluster.Namespace).Get(ctx, cluster.ChildResourceName("server"), metav1.GetOptions{})
-				Expect(err).To(Not(HaveOccurred()))
-				annotations = sts.Spec.Template.Annotations
-				return annotations
-			}, 5).Should(HaveKey("rabbitmq.com/lastRestartAt"))
-			_, err = time.Parse(time.RFC3339, annotations["rabbitmq.com/lastRestartAt"])
-			Expect(err).NotTo(HaveOccurred(), "Annotation rabbitmq.com/lastRestartAt was not a valid RFC3339 timestamp")
-
-			// delete rmq cluster
-			Expect(client.Delete(ctx, cluster)).To(Succeed())
-			waitForClusterDeletion(ctx, cluster, client)
-		},
-
-		Entry("spec.rabbitmq.additionalConfig is updated", "additional-config"),
-		Entry("spec.rabbitmq.advancedConfig is updated", "advanced-config"),
-		Entry("spec.rabbitmq.envConfig is updated", "env-config"),
-	)
 })
 
 func extractContainer(containers []corev1.Container, containerName string) corev1.Container {
@@ -1718,32 +1175,6 @@ func service(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqClust
 	return svc
 }
 
-func tlsSecretWithCACert(ctx context.Context, secretName, namespace, caCertName string) {
-	tlsData := map[string]string{
-		"tls.crt":  "this is a tls cert",
-		"tls.key":  "this is a tls key",
-		caCertName: "certificate",
-	}
-
-	_, err := createSecret(ctx, secretName, namespace, tlsData)
-
-	if !apierrors.IsAlreadyExists(err) {
-		Expect(err).NotTo(HaveOccurred())
-	}
-}
-
-func tlsSecretWithoutCACert(ctx context.Context, secretName, namespace string) {
-	tlsData := map[string]string{
-		"tls.crt": "this is a tls cert",
-		"tls.key": "this is a tls key",
-	}
-	_, err := createSecret(ctx, secretName, namespace, tlsData)
-
-	if !apierrors.IsAlreadyExists(err) {
-		Expect(err).NotTo(HaveOccurred())
-	}
-}
-
 func createSecret(ctx context.Context, secretName string, namespace string, data map[string]string) (corev1.Secret, error) {
 	secret := corev1.Secret{
 		ObjectMeta: metav1.ObjectMeta{
@@ -1755,22 +1186,6 @@ func createSecret(ctx context.Context, secretName string, namespace string, data
 
 	_, err := clientSet.CoreV1().Secrets(namespace).Create(ctx, &secret, metav1.CreateOptions{})
 	return secret, err
-}
-
-func rabbitmqClusterWithTLS(ctx context.Context, clustername string, namespace string, tlsSpec rabbitmqv1beta1.TLSSpec) *rabbitmqv1beta1.RabbitmqCluster {
-	rabbitmqCluster := &rabbitmqv1beta1.RabbitmqCluster{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      clustername,
-			Namespace: namespace,
-		},
-		Spec: rabbitmqv1beta1.RabbitmqClusterSpec{
-			TLS: tlsSpec,
-		},
-	}
-
-	Expect(client.Create(ctx, rabbitmqCluster)).To(Succeed())
-
-	return rabbitmqCluster
 }
 
 func waitForClusterCreation(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqCluster, client runtimeClient.Client) {
