+++ controllers/reconcile_finalizer.go
@@ -1,0 +1,95 @@
+package controllers
+
+import (
+	"context"
+	"fmt"
+	rabbitmqv1beta1 "github.com/rabbitmq/cluster-operator/api/v1beta1"
+	"github.com/rabbitmq/cluster-operator/internal/resource"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	clientretry "k8s.io/client-go/util/retry"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+)
+
+const deletionFinalizer = "deletion.finalizers.rabbitmqclusters.rabbitmq.com"
+
+// addFinalizerIfNeeded adds a deletion finalizer if the RabbitmqCluster does not have one yet and is not marked for deletion
+func (r *RabbitmqClusterReconciler) addFinalizerIfNeeded(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqCluster) error {
+	if rabbitmqCluster.ObjectMeta.DeletionTimestamp.IsZero() && !controllerutil.ContainsFinalizer(rabbitmqCluster, deletionFinalizer) {
+		controllerutil.AddFinalizer(rabbitmqCluster, deletionFinalizer)
+		if err := r.Client.Update(ctx, rabbitmqCluster); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (r *RabbitmqClusterReconciler) removeFinalizer(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqCluster) error {
+	controllerutil.RemoveFinalizer(rabbitmqCluster, deletionFinalizer)
+	if err := r.Client.Update(ctx, rabbitmqCluster); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (r *RabbitmqClusterReconciler) prepareForDeletion(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqCluster) error {
+	if controllerutil.ContainsFinalizer(rabbitmqCluster, deletionFinalizer) {
+		if err := clientretry.RetryOnConflict(clientretry.DefaultRetry, func() error {
+			sts := &appsv1.StatefulSet{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      rabbitmqCluster.ChildResourceName("server"),
+					Namespace: rabbitmqCluster.Namespace,
+				},
+			}
+			// Add label on all Pods to be picked up in pre-stop hook via Downward API
+			if err := r.addRabbitmqDeletionLabel(ctx, rabbitmqCluster); err != nil {
+				return fmt.Errorf("failed to add deletion markers to RabbitmqCluster Pods: %s", err.Error())
+			}
+			// Delete StatefulSet immediately after changing pod labels to minimize risk of them respawning.
+			// There is a window where the StatefulSet could respawn Pods without the deletion label in this order.
+			// But we can't delete it before because the DownwardAPI doesn't update once a Pod enters Terminating.
+			if err := r.Client.Delete(ctx, sts); client.IgnoreNotFound(err) != nil {
+				return fmt.Errorf("cannot delete StatefulSet: %s", err.Error())
+			}
+
+			return nil
+		}); err != nil {
+			r.Log.Error(err, "RabbitmqCluster deletion")
+		}
+
+		if err := r.removeFinalizer(ctx, rabbitmqCluster); err != nil {
+			r.Log.Error(err, "Failed to remove finalizer for deletion")
+			return err
+		}
+	}
+	return nil
+}
+
+func (r *RabbitmqClusterReconciler) addRabbitmqDeletionLabel(ctx context.Context, rabbitmqCluster *rabbitmqv1beta1.RabbitmqCluster) error {
+	pods := &corev1.PodList{}
+	selector, err := labels.Parse(fmt.Sprintf("app.kubernetes.io/name=%s", rabbitmqCluster.Name))
+	if err != nil {
+		return err
+	}
+	listOptions := client.ListOptions{
+		LabelSelector: selector,
+	}
+
+	if err := r.Client.List(ctx, pods, &listOptions); err != nil {
+		return err
+	}
+
+	for i := 0; i < len(pods.Items); i++ {
+		pod := &pods.Items[i]
+		pod.Labels[resource.DeletionMarker] = "true"
+		if err := r.Client.Update(ctx, pod); client.IgnoreNotFound(err) != nil {
+			return fmt.Errorf("cannot Update Pod %s in Namespace %s: %s", pod.Name, pod.Namespace, err.Error())
+		}
+	}
+
+	return nil
+}
